Incorporated feedback from FR itself

+++ inc/scatter_mj_mt.hpp
@@ -15,7 +15,7 @@
 std::set<size_t> GetRandomIndices(baby_basic &st, double norm, TRandom3 &rand3);
   
 void Process(baby_basic &st, TGraph &g, TGraph &g_full,
-	     int color, int marker, int size,
+	     int color, int marker, double size,
 	     const std::set<size_t> & indices, int nleps = 0);
 
 void Process(baby_basic &st, TGraph &g, TGraph &g_full, TH2D &h,
